refactor: 바인딩 어댑터 패키지 위치 변경 및 북마크 아이콘 리소스명 변경
