--- android/app/src/main/res/drawable/btn_bookmark.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<selector xmlns:android="http://schemas.android.com/apk/res/android">
-    <item android:state_selected="true" android:drawable="@drawable/btn_bookmark_selected" />
-
-    <item android:drawable="@drawable/btn_bookmark_unselected" />
-</selector>
